reloc: the fiducial prior loads its own marker map
